backend/requirements-llm.txt
llama-cpp-python
torch
transformers
flask
